src/extractor-service/requirements.txt
aniso8601==7.0.0
autopep8==1.4.4
Click==7.0
cycler==0.10.0
python-dotenv==0.10.3
Flask==1.0.3
Flask-RESTful==0.3.7
grpcio==1.22.0
grpcio-tools==1.22.0
itsdangerous==1.1.0
Jinja2==2.10.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
mongoengine==0.17.0
numpy==1.16.4
protobuf==3.8.0
pycodestyle==2.5.0
pymongo==3.8.0
pypaca==0.0.2
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
rope==0.14.0
selenium==3.141.0
six==1.12.0
urllib3==1.25.3
Werkzeug==0.15.4
